allow provider authorizer for http request module
