+++ .changeset/sweet-birds-watch.md
@@ -1,0 +1,5 @@
+---
+'@roadiehq/scaffolder-backend-module-http-request': patch
+---
+
+Allow configuring a custom authorizer for the backstage http request module.
