+++ plugins/scaffolder-actions/scaffolder-backend-module-http-request/src/actions/run/types.ts
@@ -24,10 +24,3 @@ export type Params = {
 export type Methods = 'GET' | 'HEAD' | 'OPTIONS' | 'POST' | 'UPDATE' | 'DELETE' | 'PUT' | 'PATCH';
 
 export type Body = string | undefined
-
-export interface HttpOptions {
-  url: string;
-  method: Methods;
-  headers: Headers;
-  body?: Body;
-}
